updated scripts

+++ A_Basic/02_Basics/01_Arithmetic_Operations/h_groceryStoreApplication.py
@@ -11,10 +11,10 @@ costOfMango = 5
 # NOTE: raw_input() takes every input as a string type only
 #       raw_input() in python 2 is input() in python 3
 
-#quantityOfApples = raw_input('quantityOfApples=')
-#quantityOfMangos = raw_input('quantityOfMangos=')
+# quantityOfApples = raw_input('quantityOfApples=')
+# quantityOfMangos = raw_input('quantityOfMangos=')
 
-quantityOfApples = float(raw_input('quantityOfApples='))
+quantityOfApples = int(raw_input('quantityOfApples='))
 quantityOfMangos = float(raw_input('quantityOfMangos='))
 
 print "type(quantityOfApples) = ", type(quantityOfApples)
